src/main/java/com/yuzf/service/impl/SellerServiceImpl.java
package com.yuzf.service.impl;

import com.yuzf.dataobject.SellerInfo;
import com.yuzf.repository.SellerInfoRepository;
import com.yuzf.service.SellerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import org.springframework.stereotype.Service;

@Service
public class SellerServiceImpl implements SellerService {

    @Autowired
    private SellerInfoRepository sellerInfoRepository;

    @Override
    public SellerInfo findSellerInfoByOpenid(String openid) {
        return sellerInfoRepository.findByOpenid(openid);
    }
}
